Only match running pods for non-job groups
This is a fix to prevent false positives when there is a group of job
pods (like a cronjob) named similarly to a deployment group.

If you have a deployment named "test" you will get pods matching

```
test-123123-123123
```

Creating a cronjob called "test-job" would then create pods matching

```
test-job-123123
```

The pods from the job would then be matched and would trigger alerts
because completed jobs are not "ready" anymore.