Fix default config merging for lone pods
A lone pod is one where the namespace itself doesn't have
"watcher: enabled" and instead a particular group of pods is enabled.

When generating the configuration for a Watch the defaults are merged
like so:

```
global config -> namespace config -> pod config
```

The bug here is that if the namespace didn't have watcher enabled, it
also didn't inherit the global defaults and would instead fallback to no
defaults. This lead to situations where a lone pod wouldn't be using the
correct defaults.