Use example.com for default development config file
Today I started receiving alerts for someone else's Kuberneretes cluster
because they used this default config as part of their deployment. I
also removed my email from the testing examples as well for a bit of
consistency.

Also snuck in a couple of other small fixes to make sure that the build
is passing again.