Cleanup old Watchers for removed workloads
This is a quick backwards compatible way to make sure Kuberwatcher
cleans up old alerts. Without this the amount of alerts will continuously
grow which can be problematic for workflows creating unique alert names.

An idea for a better more forward thinking implementation is to add
something like `metadata["kuberwatcher"]` to the alerts so that the
ownership can be clearly detected instead.